requirements.txt
GitPython==3.1.40
langchain==0.0.335
